add warning to view mode
